replaced CloverCounter in a separate file
